Fin du backed manque plus que la sécu dans bo

+++ BoManagementMS/src/main/java/iut/montreuil/BoManagementMS/app/controller/BoAnnonceController.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -16,25 +17,36 @@ import iut.montreuil.BoManagementMS.app.model.Annonce;
 @RestController
 @RequestMapping("/boannonce")
 public class BoAnnonceController {
-
+	
 	@Autowired
 	private BoAnnonceService boAnnonceService;
 	
 	@GetMapping("get")
-	public List<Annonce> getAllAnnonces() {
-		return boAnnonceService.getAllAnnonces();
+	public List<Annonce> getAllAnnonces(@RequestHeader("Authorization") String token) throws Exception {
+		if (token.equals("jwt-token-for-admin") || token.equals("jwt-token-for-normal-user")) {
+			return boAnnonceService.getAllAnnonces();	
+		} else {
+			throw new Exception("Access denied: Insufficient privileges.");
+		}
 	}
 	
 	@GetMapping("get/{id}")
-	public Optional<Annonce> getAnnonceById(@PathVariable Long id) {
-		return boAnnonceService.getAnnonceById(id);
+	public Optional<Annonce> getAnnonceById(@PathVariable Long id, @RequestHeader("Authorization") String token) throws Exception {
+		if (token.equals("jwt-token-for-admin") || token.equals("jwt-token-for-normal-user")) {
+			return boAnnonceService.getAnnonceById(id);	
+		} else {
+			throw new Exception("Access denied: Insufficient privileges.");
+		}
 	}
 	
 	@DeleteMapping
     @RequestMapping("delete/{id}")
-    public String deleteAnnonce(@PathVariable Long id) {
-		System.out.println(id);
-		boAnnonceService.deleteAnnonce(id);
-        return "Annonce "+ id +" have been deleted successfully.";
+    public String deleteAnnonce(@PathVariable Long id, @RequestHeader("Authorization") String token) {    
+        if (token.equals("jwt-token-for-admin")) {
+            boAnnonceService.deleteAnnonce(id);
+            return "Annonce " + id + " has been deleted successfully.";
+        } else {
+            return "Access denied: Insufficient privileges.";
+        }
     }
 }
